feat(port): Phases 6 and 7 -- the Bonsai UI, and a 32-bit int that only the browser could find
Done as one pass rather than two. The plan splits a read-only skeleton from an
interactive one for scheduling reasons, and building a state machine in order not
to call it is not a useful intermediate.

Five React stores become one Bonsai state machine over a single `World.t`.
`addNpc` and `clearEncounter` each write two stores in two `setState` calls
today, so there is a window where they disagree and undo covers one of them; here
there is one `World.apply` and one atomic transition. The theme stays outside,
deliberately, because it is not part of the fight and must not be undoable.

Effects are minted in exactly one place. `World.apply` is pure because ids and
timestamps are arguments, and `App_state.apply_action` is where they come from --
the id counter lives in the model, so the whole machine stays a function of the
model, the action and the clock reading that arrives as the state machine's
input.

`damageInputs` and `editingIds` become `Bonsai.assoc` per-row state. The React
version keeps them as root-level maps keyed by combatant id and prunes only
`editingIds`, so `damageInputs` leaks an entry for every combatant that ever
existed. That is not fixed here so much as made unrepresentable: there is no
root-level map to leave an entry in.

`src/App.css` and `src/index.css` are copied unchanged by a dune rule and the
class names are the React app's verbatim, so this is a logic port with an
identical-looking result.

Three bugs, all found by running it rather than by reading it.

**js_of_ocaml's `int` is 32 bits.** The build warned that 86_400_000_000_000 had
been truncated, which was my date arithmetic -- but chasing it found the real
one: JavaScript epoch milliseconds are about 1.7e12 and the wire types carried
them as `int`. Every native test passed and the browser would have been silently
wrong. They are `float` now, which is what a JSON number is anyway.

**A `Bonsai.state` setter takes the new value, so it closes over the old one.**
Ticking four player characters put one in the fight: each handler read the same
stale set and the last write won. The selection is a `state_machine0` now, where
an action says what to do rather than what the answer is. This is a good argument
for state machines over states and it is in the code where it happened.

**The difficulty line printed its label twice**, because the readout wrote a
"Difficulty: <label>" heading and then called `to_string_hum`, which begins with
the label. The readout composes from the fields; `to_string_hum` is for the CLI.

Two things are polled rather than edge-triggered, for the same reason: saving and
the difficulty analysis. A burst of keystrokes should cost one write and one
analysis, not one of each per character. Polling also removes the plan's monotone
request counter -- there is only ever one analysis in flight -- which is the
better kind of simplification.

The exact solver is a few tens of millions of flops, so it never runs on the
render path. The signature it debounces against is built from the fields the
model reads and so, deliberately, not from names or notes: typing a note cannot
cost an analysis because the note is not in the value being compared.

Verified in a browser end to end: the four shipped characters load with their
defence targets (2, 7, 10, 10), the Robber preset fills the form with the same
numbers the CLI prints, three NPCs get numbered 1 to 3, the readout says
"Trivial -- 99% win, expect 1.3 casualties, median 9 rounds" over 5,427 states --
matching the expect test exactly -- and a reload restores everything from
`sct.v2.app`.

+++ lib/symbaroum/codec/wire_v1.ml
@@ -164,7 +164,7 @@ module Bestiary_entry = struct
     ; armor : string
     ; pain_threshold : int option
     ; note : string
-    ; updated_at_ms : int
+    ; updated_at_ms : float
     }
   [@@deriving compare, equal, sexp_of]
 
@@ -178,7 +178,7 @@ module Bestiary_entry = struct
     and armor = D.field_or "armor" D.string ~default:""
     and pain_threshold = D.field_opt "painThreshold" D.int
     and note = D.field_or "note" D.string ~default:""
-    and updated_at_ms = D.field_or "updatedAt" D.int ~default:0 in
+    and updated_at_ms = D.field_or "updatedAt" D.float ~default:0. in
     { id
     ; monster_type
     ; initiative
@@ -195,7 +195,7 @@ end
 module History_entry = struct
   type t =
     { id : string
-    ; timestamp_ms : int
+    ; timestamp_ms : float
     ; label : string
     ; encounter : Encounter.t
     }
@@ -204,7 +204,7 @@ module History_entry = struct
   let decoder =
     let open D.Let_syntax in
     let%map id = D.field "id" D.string
-    and timestamp_ms = D.field_or "timestamp" D.int ~default:0
+    and timestamp_ms = D.field_or "timestamp" D.float ~default:0.
     and label = D.field_or "label" D.string ~default:""
     and encounter = D.field "encounter" Encounter.decoder in
     { id; timestamp_ms; label; encounter }
